Attempted Campfire, but wierdness fought back.

+++ src/main/java/net/einsteinsci/betterbeginnings/register/recipe/CampfirePotRecipes.java
@@ -96,6 +96,6 @@ public class CampfirePotRecipes
 			return stack.getItem().getSmeltingExperience(stack);
 		}
 
-		return ((Float)entry.getValue()).floatValue();
+		return (Float)entry.getValue();
 	}
 }
